// src/app/shared/interfaces/metacolumn.interfaces.ts
// Interfaz TypeScript llamada MetaDataColumn para representar la estructura de los metadatos de las columnas de una tabla
export interface MetaDataColumn {
  // Propiedad 'field' que almacena el nombre del campo asociado a la columna
  field: string;
  // Propiedad 'title' que almacena el título o encabezado asociado a la columna
  title: string;
}